forbid trailing '/' in password names
dirname strips them, which is expected, but then age is unable to save
file due to missing parent directory.

before switching to dirname, pa used to create ".age" file under the
password name, which is probably a mistake, and it's unclear what the
user intent would be there, so best to just avoid it altogether.

also, reword 'Category' -> 'Name' because category is part of the name
and the error sounds clearer this way: the user may not know that they
create a "category" by typing an errant slash.